Massive optimisation update
Changed the script from requests to httpx, moved the api calls to async, to allow for much more efficient calls. Additionally, the required api calls for a system is much more procedural, as it is dependent on finding whether a system is colonisable rather than the original approach of finding all info for all systems.